packaging cooperation management,store, components, models, validators, reworked dependencies
