+++ src/cooperation/components/CooperationInfo.vue
@@ -67,9 +67,9 @@ import { mapGetters } from 'vuex';
 import Button from 'primevue/button';
 import Dialog from 'primevue/dialog';
 import Breadcrumb from '@/components/Breadcrumb.vue';
-import EditCooperationForm from '@/components/EditCooperationInfo.vue';
+import EditCooperationForm from '@/cooperation/components/EditCooperationForm.vue';
 import ManageHouses from '@/houses/components/ManageHouses.vue';
-import { CooperationActionEnum, CooperationGettersEnum } from '@/store/cooperation/types';
+import { CooperationActionEnum, CooperationGettersEnum } from '@/cooperation/store/types';
 import { StoreModuleEnum } from '@/store/types';
 
 export default defineComponent({
