+++ src/cooperation/components/RegisterCooperation.vue
@@ -59,7 +59,7 @@ import {
 import InputText from 'primevue/inputtext';
 import Button from 'primevue/button';
 import { StoreModuleEnum } from '@/store/types';
-import { CooperationActionEnum } from '@/store/cooperation/types';
+import { CooperationActionEnum } from '@/cooperation/store/types';
 import { RoutesEnum } from '@/router/types';
 
 export default defineComponent({
