+++ src/cooperation/models/put-cooperationDTO.model.ts
@@ -1,16 +1,13 @@
-import { CooperationContactsInterface } from '@/store/cooperation/types';
+import { CooperationContactsInterface } from '@/cooperation/store/types';
 import { AddressDTOModel } from '@/houses/models/addressDTO.model';
-import { CooperationModel } from './cooperation.model';
-import { HouseDTOModel } from '@/houses/models/houseDTO.model';
-import { HouseModel } from '@/houses/models/house.model';
+import { CooperationModel } from '@/cooperation/models/cooperation.model';
 
-export class CooperationDTOModel {
+export class CooperationPutDTOModel {
   public id: number;
   public name: string;
   public usreo: string;
   public iban: string;
   public address: AddressDTOModel;
-  public houses: Array<HouseDTOModel>;
   public contacts: Array<CooperationContactsInterface>;
 
   constructor(data: CooperationModel) {
@@ -19,7 +16,6 @@ export class CooperationDTOModel {
     this.usreo = data.edrpou;
     this.iban = data.iban;
     this.address = new AddressDTOModel(data.address);
-    this.houses = data.houses.map((el: HouseModel) => new HouseDTOModel(el));
     this.contacts = data.contacts;
   }
 }
