+++ src/cooperation/store/getters.ts
@@ -1,7 +1,7 @@
 import { GetterTree } from 'vuex';
 
 import { RootStateInterface } from '@/store/types';
-import { CooperationStateInterface, Getters } from '@/store/cooperation/types';
+import { CooperationStateInterface, Getters } from '@/cooperation/store/types';
 
 export const getters: GetterTree<CooperationStateInterface, RootStateInterface> & Getters = {
   getCooperationNameAndEdrpou: (state) => {
