+++ src/store/index.ts
@@ -1,7 +1,7 @@
 import { createStore } from 'vuex';
 import { RootStateInterface } from '@/store/types';
 import { authorizationStore } from '@/store/authorization';
-import { cooperationStore } from '@/store/cooperation';
+import { cooperationStore } from '@/cooperation/store';
 import { localStorageStore } from '@/store/localstorage';
 import { housesStore } from '@/houses/store';
 import { userStore } from '@/store/user';
