+++ src/store/types.ts
@@ -1,7 +1,7 @@
 import { AxiosError, AxiosResponse } from 'axios';
 import { UserStateInterface } from '@/store/user/types';
 import { AuthorizationStateInterface } from '@/store/authorization/types';
-import { CooperationStateInterface } from '@/store/cooperation/types';
+import { CooperationStateInterface } from '@/cooperation/store/types';
 import { LocalStorageStateInterface } from '@/store/localstorage/types';
 import { HousesStateInterface } from '@/houses/store/types';
 import { PollsStateInterface } from '@/store/polls/types';
